updating typings

+++ src/commands/admin_cog.py
@@ -4,6 +4,8 @@ import discord
 from discord.commands import SlashCommandGroup
 from discord.ext import commands
 
+from src.models import GuildDTO
+
 logger = logging.getLogger(__name__)
 logger.setLevel(level=logging.DEBUG)
 ch = logging.StreamHandler()
@@ -13,10 +15,8 @@ logger.addHandler(ch)
 
 
 from src import BansheeBot
-
 from src.views.admin_views import AdminRoleSelectView
 from src.views.guild_views import GuildViews
-
 from src.raiderIO import RaiderIOClient
 
 
@@ -28,7 +28,7 @@ class Admin(commands.Cog):
     admin = SlashCommandGroup(name="admin", description="Admin commands")
 
     @admin.command(name="set_role", description="Set the role you would like to track")
-    @discord.ext.commands.has_permissions(administrator=True)
+    @commands.has_permissions(administrator=True)
     async def set_role(self, ctx: discord.ApplicationContext):
         view = AdminRoleSelectView()
         await ctx.respond("Select roles: ", view=view, ephemeral=True)
@@ -37,11 +37,15 @@ class Admin(commands.Cog):
         name="delete_last_messages",
         description="Delete the last n messages in the current channel",
     )
-    @discord.ext.commands.has_permissions(administrator=True)
+    @commands.has_permissions(administrator=True)
     async def delete_last_messages(
         self, ctx: discord.ApplicationContext, amount: int
     ) -> None:
-        messages = await ctx.channel.history(limit=amount).flatten()
+
+        if ctx.channel is None:
+            raise Exception("Channel was none")
+
+        messages = await ctx.channel.history(limit=amount).flatten()  # type: ignore
         for msg in messages:
             await msg.delete()
 
@@ -51,12 +55,12 @@ class Admin(commands.Cog):
         name="set_wow_guild",
         description="Set this servers World of Warcraft guild",
     )
-    @discord.ext.commands.has_permissions(administrator=True)
+    @commands.has_permissions(administrator=True)
     async def set_wow_guild(
         self,
         ctx: discord.ApplicationContext,
         name: str,
-        realm: Optional[str] = "Dalaran",
+        realm: str = "Dalaran",
     ):
         try:
             guild_io = await RaiderIOClient.getGuildProfile(name, realm)
@@ -64,7 +68,8 @@ class Admin(commands.Cog):
             if guild_io is None:
                 await ctx.respond(f"Guild {name}-{realm} was not found.")
             else:
-
+                if ctx.guild is None:
+                    raise Exception("Guild was none")
                 # add new guild to database and link it to current server
                 wow_guild = await self.bot.db.addWowGuild(ctx.guild.id, guild_io)
 
@@ -81,12 +86,12 @@ class Admin(commands.Cog):
         name="add_character_to_guild",
         description="Adds a WoW character to the server's WoW guild",
     )
-    @discord.ext.commands.has_permissions(administrator=True)
+    @commands.has_permissions(administrator=True)
     async def add_character_to_guild(
         self,
         ctx: discord.ApplicationContext,
         name: str,
-        realm: Optional[str] = "Dalaran",
+        realm: str = "Dalaran",
     ):
         try:
             character_io = await RaiderIOClient.getCharacterProfile(name, realm)
@@ -96,16 +101,20 @@ class Admin(commands.Cog):
                 )
                 return
             else:
+
+                if ctx.guild is None:
+                    raise Exception("Guild was none")
+
                 wow_character = await self.bot.db.addWowCharacterToWowGuild(
                     ctx.guild.id, character_io, ctx.author.id
                 )
 
-                if wow_character is None:
+                if wow_character is None or wow_character.wow_guild is None:
                     await ctx.respond("There was a problem adding chracter")
                     raise Exception
                 else:
                     await ctx.respond(
-                        f"{wow_character.wow_character_name} was added to {wow_character.wow_guild.wow_guild_name}"
+                        f"{wow_character.name} was added to {wow_character.wow_guild.name}"
                     )
 
         except Exception as err:
@@ -116,16 +125,21 @@ class Admin(commands.Cog):
         name="get_guild_summary",
         description="Get a summary of the current guild",
     )
-    @discord.ext.commands.has_permissions(administrator=True)
+    @commands.has_permissions(administrator=True)
     async def get_guild_summary(self, ctx: discord.ApplicationContext):
         try:
-            wow_guild = await self.bot.db.getWowGuild(ctx.guild.id)
+            if ctx.guild is None:
+                raise Exception("Guild was none")
 
-            if wow_guild is None or len(wow_guild.wow_characters) == 0:
+            wow_guild_db = await self.bot.db.getWowGuild(ctx.guild.id)
+
+            if wow_guild_db is None or len(wow_guild_db.wow_characters) == 0:
                 print("Wow guild was none")
                 raise Exception
 
-            embed = GuildViews.getGuildSummary(wow_guild, wow_guild.wow_characters)
+            wow_guild = GuildDTO(**wow_guild_db.model_dump())
+
+            embed = GuildViews.getGuildSummary(wow_guild, wow_guild.characters)
 
             await ctx.respond(embed=embed)
 
@@ -134,6 +148,6 @@ class Admin(commands.Cog):
             await ctx.respond("Something went wrong")
 
 
-def setup(bot: commands.Bot) -> None:
+def setup(bot: BansheeBot) -> None:
     bot.add_cog(Admin(bot))
     logger.info("Admin cog has loaded successfully")
